feature : 검색 결과 페이지 / 강의 검색카드 PC,태블릿 반응형 UI 구현 "#81"
